WIP on email parsing in E2E tests
This isn't ready for merging quite yet, because a couple changes will
need to be made. Most importantly, there's a timing problem. In the
sample test spec, console.log() gets called during the callback, *after*
the test case (the it() function) has returned! If you put an expect()
in there, it won't actually be able to fail.

The solution is to do something with promises, and fulfill the promise
instead of calling a callback -- and then the Protractor version of
Jasmine's expect() will wait for the promise object to be fulfilled. The
next commit will contain that solution; this commit is to save this code
which, even if it's not useful for testing, at least works and can be
used as a jumping-off point.