WIP - email parsing functions now return promises
There's a trick to using these functions in unit tests, and that's to
follow the pattern found in the sample test -- call browser.wait() with
the function as the first parameter. Since the function returns a
promise, and browser.wait(), if it gets a function returning a promise,
will wait for that promise to be fulfilled, we get precisely what we
wanted: the test waits for the promise to be fulfilled before moving on,
and the expect() has a fulfilled promise to work with.