Make all E2E tests work again
To make all E2E tests work, we may have to disable the afterEach and
beforeEach functions that depend on Angular being present. (E.g.,
test/app/allspecs/e2e/beforeEach.spec.js for one.) There may be a better
way of doing this -- setting up beforeEach and afterEach only for test
functions that match a certain pattern, or whatever -- but so far a
quick bit of research hasn't turned up a better way. If we do find a
better way, where beforeEach and afterEach functions can be set up only
on certain tests but not all, we can do it right -- but for now, just
disable the beforeEach and afterEach until such a time as we can find
the *right* way to do this.