Add non-working "forgot password" test (disabled)
The test is disabled by renaming the .spec.js file so it won't be picked
up in running all tests. Checking it in, despite the fact that it's not
working right now, because it's a good start. Perhaps at a later date
we'll be able to revisit it and fix it to work *properly*.